Refactor stand-alone AngularCompiler, which will wrap the overall compilation pipeline.
PiperOrigin-RevId: 226068364
