+++ angular/lib/src/source_gen/common/ng_compiler.dart
@@ -1,4 +1,5 @@
 import 'package:build/build.dart' hide AssetReader;
+import 'package:angular/src/compiler/angular_compiler.dart';
 import 'package:angular/src/compiler/ast_directive_normalizer.dart';
 import 'package:angular/src/compiler/ast_template_parser.dart';
 import 'package:angular/src/compiler/expression_parser/lexer.dart' as ng;
@@ -7,23 +8,36 @@ import 'package:angular/src/compiler/offline_compiler.dart';
 import 'package:angular/src/compiler/output/dart_emitter.dart';
 import 'package:angular/src/compiler/schema/dom_element_schema_registry.dart';
 import 'package:angular/src/compiler/stylesheet_compiler/style_compiler.dart';
+import 'package:angular/src/compiler/view_compiler/directive_compiler.dart';
 import 'package:angular/src/compiler/view_compiler/view_compiler.dart';
 import 'package:angular_compiler/angular_compiler.dart';
 import 'package:angular_compiler/cli.dart';
 
-OfflineCompiler createTemplateCompiler(
+AngularCompiler createAngularCompiler(
+    BuildStep buildStep, CompilerFlags flags) {
+  final schemaRegistry = DomElementSchemaRegistry();
+  final parser = ng.Parser(ng.Lexer());
+  return AngularCompiler(
+      OfflineCompiler(
+        DirectiveCompiler(schemaRegistry),
+        StyleCompiler(flags),
+        ViewCompiler(flags, parser, schemaRegistry),
+        DartEmitter(),
+        {},
+      ),
+      AstDirectiveNormalizer(NgAssetReader.fromBuildStep(buildStep)),
+      AstTemplateParser(schemaRegistry, parser, flags));
+}
+
+OfflineCompiler createOfflineCompiler(
   BuildStep buildStep,
   CompilerFlags flags,
 ) {
-  final reader = NgAssetReader.fromBuildStep(buildStep);
-  final parser = ng.Parser(ng.Lexer());
   final schemaRegistry = DomElementSchemaRegistry();
-  final templateParser = AstTemplateParser(schemaRegistry, parser, flags);
   return OfflineCompiler(
-    AstDirectiveNormalizer(reader),
-    templateParser,
+    DirectiveCompiler(schemaRegistry),
     StyleCompiler(flags),
-    ViewCompiler(flags, parser, schemaRegistry),
+    ViewCompiler(flags, ng.Parser(ng.Lexer()), schemaRegistry),
     DartEmitter(),
     {},
   );
